+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 ---
 
 ## master
-* Support Spies code generation
+* Support Spies code generation ( suspend functions are not supported )
 * Adding test to cover code generation regression using Ksp
 
 ## 2.8.1